helper method added.
